Extract Entry from Indexer
